Bug 714874 - Fix Java warnings about redundant casts. r=pcwalton r=dougt a=dougt
UltraBlame original commit: 16f5cf776059a57038a46b9868819c9edddda331

+++ mobile/android/base/gfx/ScrollbarLayer.java
@@ -152,7 +152,7 @@ public class ScrollbarLayer extends TileLayer {
         foregroundPaint.setAntiAlias(true);
         foregroundPaint.setStyle(Paint.Style.FILL);
         
-        foregroundPaint.setColor(Color.argb((int)Math.round(mOpacity * 127), 0, 0, 0));
+        foregroundPaint.setColor(Color.argb(Math.round(mOpacity * 127), 0, 0, 0));
 
         mCanvas.drawColor(Color.argb(0, 0, 0, 0), PorterDuff.Mode.CLEAR);
         mCanvas.drawCircle(CAP_RADIUS, CAP_RADIUS, CAP_RADIUS, foregroundPaint);
